renamed slack_module.py to slack_addon.py
